+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 - Fetch users by handle.
 - Add friend by handle.
 - Filter leaderboards by ID.
+- Add storage permission when storing data.
 
 ### Fixed
 - Correctly handle callbacks when sending match data.